implemented the feedback from Gabriel
The feedback: First
The form is not consistent with the design of the rest of the app. Make a separate page that opens for the form. The current one is working with Javascript. What happens if I turn the Javascript off? (it won't work)

Second
The form does not give proper errors and continues to submit even if the submission itself is invalid. The only errors are the default html form ones instead of the Laravel and Javascript ones.